Moved synth_server to separate file.
