+++ psonic/psonic.py
@@ -1,25 +1,15 @@
 import random
-import time
-#TODO: Optimize imports
 from .samples import Sample
-from .synthesizers import BEEP, SAW
+from .synthesizers import SAW
 from .notes import C5
 from .internals.chords import _CHORD_QUALITY
 from .internals.scales import _SCALE_MODE
-
-from pythonosc import osc_message_builder  # osc support
-from pythonosc import udp_client
+from .synth_server import (
+    SonicPi,
+    use_synth,
+)
 
 __debug = False
-
-## Module attributes ##
-_current_synth = BEEP
-
-## Module methodes ##
-def use_synth(synth):
-    global _current_synth
-    _current_synth = synth
-
 
 def synth(name, note=None, attack=None, decay=None,
     sustain_level=None, sustain=None, release=None,
@@ -35,7 +25,6 @@ def synth(name, note=None, attack=None, decay=None,
     _debug('synth command={}'.format(command))
     synth_server.synth(command)
 
-
 def play(note, attack=None, decay=None,
     sustain_level=None, sustain=None, release=None,
     cutoff=None, cutoff_attack=None, amp=None, pan=None):
@@ -49,7 +38,6 @@ def play(note, attack=None, decay=None,
 
     _debug('play command={}'.format(command))
     synth_server.play(command)
-
 
 def play_pattern_timed(notes, times, release=None):
     """play notes
@@ -65,13 +53,11 @@ def play_pattern_timed(notes, times, release=None):
             play(i, release=release)
             sleep(t)
 
-
 def play_pattern(notes):
     """:param notes:
     :return:
     """
     play_pattern_timed(notes, 1)
-
 
 def sample(sample, rate=None, attack=None, sustain=None,
 	   release=None, beat_stretch=None, start=None,
@@ -91,7 +77,6 @@ def sample(sample, rate=None, attack=None, sustain=None,
     _debug('sample command={}'.format(command))
     synth_server.sample(command)
 
-
 def sleep(duration):
     """the same as time.sleep
     :param duration:
@@ -107,14 +92,12 @@ def sample_duration(sample):
     """
     return sample.duration
 
-
 def one_in(max):
     """random function  returns True in one of max cases
     :param max:
     :return: boolean
     """
     return random.randint(1, max) == 1
-
 
 def chord(root_note, chord_quality):
     """Generates a list of notes of a chord
@@ -132,7 +115,6 @@ def chord(root_note, chord_quality):
         result.append(n)
 
     return result
-
 
 def scale(root_note, scale_mode, num_octaves=1):
     """Genarates a liste of notes of scale
@@ -164,108 +146,11 @@ def stop():
 def send_message(message, *parameter):
     synth_server.send_message(message, *parameter)
 
-## Compound classes ##
-class SonicPiCommon:
-
-    UDP_IP = "127.0.0.1"
-
-    def send(self, command):
-        pass
-
-    def sleep(self, duration):
-        time.sleep(duration)
-
-    def sample(self, command):
-        self.send(command)
-
-
-## Connection classes ##
-class SonicPi(SonicPiCommon):
-    """Communiction to Sonic Pi"""
-
-    UDP_PORT = 4557
-    UDP_PORT_OSC_MESSAGE = 4559
-    GUI_ID = 'SONIC_PI_PYTHON'
-
-    RUN_COMMAND = "/run-code"
-    STOP_COMMAND = "/stop-all-jobs"
-
-    def __init__(self):
-        self.client = udp_client.UDPClient(
-            SonicPi.UDP_IP,
-            SonicPi.UDP_PORT
-        )
-        self.client_for_messages = udp_client.UDPClient(
-            self.UDP_IP,
-            self.UDP_PORT_OSC_MESSAGE
-        )
-
-    def play(self, command):
-        command = 'use_synth :{0}\n'.format(_current_synth.name) + command
-        self.send(command)
-
-    def synth(self, command):
-        self.send(command)
-
-    def send(self, command):
-        self.send_command(SonicPi.RUN_COMMAND, command)
-
-    def stop(self):
-        self.send_command(SonicPi.STOP_COMMAND)
-
-    def test_connection(self):
-        # OSC::Server.new(PORT)
-        # abort("ERROR: Sonic Pi is not listening on #{PORT} - is it running?")
-        pass
-
-    def send_command(self, address, argument=''):
-        msg = osc_message_builder.OscMessageBuilder(address=address)
-        msg.add_arg('SONIC_PI_PYTHON')
-        if argument != "":
-            msg.add_arg(argument)
-        msg = msg.build()
-        self.client.send(msg)
-
-    def send_message(self,message, *parameters):
-        msg = osc_message_builder.OscMessageBuilder(message)
-        for p in parameters:
-            msg.add_arg(p)
-        msg = msg.build()
-        self.client_for_messages.send(msg)
-
-class SonicPiNew(SonicPiCommon):
-    """Communiction to Sonic Pi"""
-
-    UDP_PORT = 4559
-
-    def __init__(self):
-        self.client = udp_client.UDPClient(
-            self.UDP_IP,
-            self.UDP_PORT
-        )
-        self.commandServer = SonicPi()
-        # x= 'live_loop :py do\n  nv=sync "/SENDOSC"\n  puts nv\n  eval(nv[0])\nend'
-        # self.commandServer.run(x)
-
-    def set_OSC_receiver(self, source):
-        self.commandServer.run(source)
-
-    def send(self, address, *message):
-        msg = osc_message_builder.OscMessageBuilder(address)
-        for m in message:
-            msg.add_arg(m)
-        msg = msg.build()
-        self.client.send(msg)
-
-    def play(self, command):
-        self.send(command)
-
 
 synth_server = SonicPi()
 
-## system functions ##
-def _debug(*allargs):  # simple debug function for working in different environments
-    if __debug: print(allargs)
+def _debug(*args):
+    if __debug: print(args)
 
 
 if __name__ == '__main__':
